added client and server dirs. commands top level. additional refactor
