+++ client/libp2p/message_sender.go
@@ -7,7 +7,7 @@ import (
 	"time"
 
 	pb "github.com/filecoin-project/storetheindex/api/v1/finder/pb"
-	"github.com/filecoin-project/storetheindex/internal/net"
+	"github.com/filecoin-project/storetheindex/server/net"
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/network"
 	"github.com/libp2p/go-libp2p-core/peer"
