+++ command/synthetic.go
