+++ server/finder/libp2p/server.go
@@ -8,8 +8,8 @@ import (
 	"github.com/filecoin-project/go-indexer-core"
 	pb "github.com/filecoin-project/storetheindex/api/v1/finder/pb"
 	"github.com/filecoin-project/storetheindex/internal/finder"
-	"github.com/filecoin-project/storetheindex/internal/net"
-	p2phandler "github.com/filecoin-project/storetheindex/resources/handler/libp2p"
+	p2phandler "github.com/filecoin-project/storetheindex/server/handler/libp2p"
+	"github.com/filecoin-project/storetheindex/server/net"
 	logging "github.com/ipfs/go-log/v2"
 	"github.com/libp2p/go-libp2p-core/host"
 	"github.com/libp2p/go-libp2p-core/network"
